changed the tests, most are failing, compared to what there was before I HOPE that I just inserted comments and couple of more tests at the end :)
